simplified code, added tests
