--- app/reports/mi-report/index.js
@@ -1,5 +1,0 @@
-const { createMIReport } = require('./create')
-
-module.exports = {
-  createMIReport
-}
