+++ app/reports/report-processing/get-currency.js
