--- app/reports/suppressed-report/index.js
@@ -1,5 +1,0 @@
-const { createSuppressedReport } = require('./create')
-
-module.exports = {
-  createSuppressedReport
-}
