+++ test/integration/local/create-mi-report.test.js
@@ -4,7 +4,7 @@ const { PAYMENT_EVENT } = require('../../../app/constants/event-types')
 
 const { reportsConfig } = require('../../../app/config')
 const { getBlob } = require('../../helpers/get-blob')
-const { createMIReport } = require('../../../app/reports/mi-report')
+const { createMIReport } = require('../../../app/reports/create-mi-report')
 const { initialise, getClient } = require('../../../app/storage')
 
 let paymentClient
