+++ test/integration/narrow/create-mi-report.test.js
@@ -9,7 +9,7 @@ const mockTableClient = {
 
 const { reportsConfig } = require('../../../app/config')
 const { stringifyEventData } = require('../../helpers/stringify-event-data')
-const { createMIReport } = require('../../../app/reports/mi-report')
+const { createMIReport } = require('../../../app/reports/create-mi-report')
 
 let extractedEvent
 let enrichedEvent
@@ -19,9 +19,9 @@ let events
 describe('create mi report', () => {
   beforeEach(() => {
     jest.clearAllMocks()
-    extractedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/extracted')))
-    enrichedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/enriched')))
-    resetEvent = JSON.parse(JSON.stringify(require('../../mocks/events/reset')))
+    extractedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/mi/extracted')))
+    enrichedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/mi/enriched')))
+    resetEvent = JSON.parse(JSON.stringify(require('../../mocks/events/mi/reset')))
 
     stringifyEventData(extractedEvent)
     stringifyEventData(enrichedEvent)
