+++ test/mocks/events/mi/submitted.js
@@ -1,4 +1,4 @@
-const { PAYMENT_SUBMITTED } = require('../../../app/constants/events')
+const { PAYMENT_SUBMITTED } = require('../../../../app/constants/events')
 const event = require('./event')
 
 module.exports = {
