+++ test/mocks/events/suppressed/enriched.js
@@ -1,0 +1,7 @@
+const { PAYMENT_ENRICHED } = require('../../../../app/constants/events')
+const event = require('./event')
+
+module.exports = {
+  ...event,
+  type: PAYMENT_ENRICHED
+}
