+++ test/mocks/events/suppressed/grouped-event.js
@@ -1,0 +1,7 @@
+const { CORRELATION_ID } = require('../../values/correlation-id')
+const event = require('./event')
+
+module.exports = {
+  correlationId: CORRELATION_ID,
+  events: [event]
+}
