+++ test/mocks/events/suppressed/suppressed.js
@@ -1,0 +1,7 @@
+const { PAYMENT_SUPPRESSED } = require('../../../../app/constants/events')
+const event = require('./event')
+
+module.exports = {
+  ...event,
+  type: PAYMENT_SUPPRESSED
+}
