+++ test/mocks/values/credit-ap.js
@@ -1,0 +1,3 @@
+module.exports = {
+  CREDIT_AP: -50000
+}
