+++ test/mocks/values/delta-value.js
@@ -1,0 +1,3 @@
+module.exports = {
+  DELTA_VALUE: -50000
+}
