+++ test/mocks/values/suppressed-ar.js
@@ -1,0 +1,3 @@
+module.exports = {
+  SUPPRESSED_AR: 50000
+}
