+++ test/unit/reports/create-suppressed-report.test.js
@@ -17,15 +17,15 @@ jest.mock('../../../app/storage')
 const { writeFile: mockWriteFile } = require('../../../app/storage')
 
 const { reportsConfig } = require('../../../app/config')
-const { createMIReport } = require('../../../app/reports/create-mi-report')
+const { createSuppressedReport } = require('../../../app/reports/create-suppressed-report')
 
-const event = require('../../mocks/events/mi/event')
-const groupedEvent = require('../../mocks/events/mi/grouped-event')
-const reportLine = require('../../mocks/mi-report-line')
+const event = require('../../mocks/events/suppressed/suppressed')
+const groupedEvent = require('../../mocks/events/suppressed/grouped-event')
+const reportLine = require('../../mocks/suppressed-report-line')
 const csv = require('../../mocks/csv')
-const { MI_REPORT } = require('../../../app/constants/report-types')
+const { SUPPRESSED_REPORT } = require('../../../app/constants/report-types')
 
-describe('create mi report', () => {
+describe('create suppressed report', () => {
   beforeEach(() => {
     jest.clearAllMocks()
 
@@ -37,69 +37,69 @@ describe('create mi report', () => {
   })
 
   test('should get all payment events', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockGetEvents).toHaveBeenCalledTimes(1)
   })
 
   test('should group events by correlation id', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockGroupEventsByCorrelationId).toHaveBeenCalledWith([event])
   })
 
   test('should group events by correlation id once', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockGroupEventsByCorrelationId).toHaveBeenCalledTimes(1)
   })
 
   test('should order grouped events', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockOrderGroupedEvents).toHaveBeenCalledWith([groupedEvent])
   })
 
   test('should order grouped events once', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockOrderGroupedEvents).toHaveBeenCalledTimes(1)
   })
 
   test('should get report lines', async () => {
-    await createMIReport()
-    expect(mockGetReportLines).toHaveBeenCalledWith([groupedEvent], MI_REPORT)
+    await createSuppressedReport()
+    expect(mockGetReportLines).toHaveBeenCalledWith([groupedEvent], SUPPRESSED_REPORT)
   })
 
   test('should get report lines once', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockGetReportLines).toHaveBeenCalledTimes(1)
   })
 
   test('should convert report lines to csv if report lines', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockConvertToCSV).toHaveBeenCalledWith([reportLine])
   })
 
   test('should convert report lines to csv once if report lines', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockConvertToCSV).toHaveBeenCalledTimes(1)
   })
 
   test('should not convert report lines to csv if no report lines', async () => {
     mockGetReportLines.mockReturnValue([])
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockConvertToCSV).not.toHaveBeenCalled()
   })
 
   test('should write csv to MI report file if report lines', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockWriteFile).toHaveBeenCalledWith(reportsConfig.miReportName, csv)
   })
 
   test('should write csv to MI report file once if report lines', async () => {
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockWriteFile).toHaveBeenCalledTimes(1)
   })
 
   test('should not write csv to MI report file if no report lines', async () => {
     mockGetReportLines.mockReturnValue([])
-    await createMIReport()
+    await createSuppressedReport()
     expect(mockWriteFile).not.toHaveBeenCalled()
   })
 })
