+++ test/unit/reports/report-processing/get-events.test.js
@@ -1,5 +1,5 @@
-jest.mock('../../../app/storage')
-const { getClient: mockGetClient, odata: mockOdata } = require('../../../app/storage')
+jest.mock('../../../../app/storage')
+const { getClient: mockGetClient, odata: mockOdata } = require('../../../../app/storage')
 
 const mockListEntities = jest.fn()
 
@@ -7,11 +7,11 @@ const mockTableClient = {
   listEntities: mockListEntities
 }
 
-const { PAYMENT_EVENT } = require('../../../app/constants/event-types')
+const { PAYMENT_EVENT } = require('../../../../app/constants/event-types')
 
-const { stringifyEventData } = require('../../helpers/stringify-event-data')
+const { stringifyEventData } = require('../../../helpers/stringify-event-data')
 
-const { getEvents } = require('../../../app/reports/shared/get-events')
+const { getEvents } = require('../../../../app/reports/report-processing/get-events')
 
 let extractedEvent
 let enrichedEvent
@@ -21,8 +21,8 @@ describe('get events', () => {
   beforeEach(() => {
     jest.clearAllMocks()
 
-    extractedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/extracted')))
-    enrichedEvent = JSON.parse(JSON.stringify(require('../../mocks/events/enriched')))
+    extractedEvent = JSON.parse(JSON.stringify(require('../../../mocks/events/mi/extracted')))
+    enrichedEvent = JSON.parse(JSON.stringify(require('../../../mocks/events/mi/enriched')))
 
     stringifyEventData(extractedEvent)
     stringifyEventData(enrichedEvent)
